refector: departament organize package
