+++ src/main/java/com/rgv04/hr/domain/dto/EmployeeManagerModel.java
@@ -6,8 +6,6 @@ import java.time.OffsetDateTime;
 import org.springframework.hateoas.RepresentationModel;
 import org.springframework.hateoas.server.core.Relation;
 
-import com.rgv04.hr.domain.departament.DepartamentModel;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
