+++ src/main/java/com/rgv04/hr/domain/dto/EmployeeModel.java
@@ -6,7 +6,6 @@ import java.time.OffsetDateTime;
 import org.springframework.hateoas.RepresentationModel;
 import org.springframework.hateoas.server.core.Relation;
 
-import com.rgv04.hr.domain.departament.DepartamentModel;
 import com.rgv04.hr.domain.job.controller.model.JobModel;
 
 import lombok.AllArgsConstructor;
