+++ src/main/java/com/rgv04/hr/domain/jobhistory/JobHistoryModel.java
@@ -5,11 +5,12 @@ import java.time.OffsetDateTime;
 import org.springframework.hateoas.RepresentationModel;
 import org.springframework.hateoas.server.core.Relation;
 
-import com.rgv04.hr.domain.departament.DepartamentModel;
+import com.rgv04.hr.domain.dto.DepartamentModel;
 import com.rgv04.hr.domain.job.controller.model.JobModel;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -19,6 +20,7 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
+@EqualsAndHashCode(callSuper = false)
 @Relation(collectionRelation = "jobHistories")
 public class JobHistoryModel extends RepresentationModel<JobHistoryModel> {
