+++ src/main/java/com/rgv04/hr/domain/model/Departament.java
@@ -1,4 +1,4 @@
-package com.rgv04.hr.domain.departament;
+package com.rgv04.hr.domain.model;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -6,8 +6,6 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
-
-import com.rgv04.hr.domain.model.Location;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
