+++ src/main/java/com/rgv04/hr/domain/repository/DepartamentRepository.java
@@ -1,9 +1,11 @@
-package com.rgv04.hr.domain.departament;
+package com.rgv04.hr.domain.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+
+import com.rgv04.hr.domain.model.Departament;
 
 public interface DepartamentRepository extends JpaRepository<Departament, Long> {
